Play flipbooks and particle sheets at exact variable frame durations
A flipbook source such as a GIF can now state an exact duration per frame.
The texture catalogue and texture table carry those durations with
cumulative ends, and particle rendering uploads one bounded table of distinct
timelines per world so the step shader picks each particle's cell by age.
Fixed-rate sheets keep offset zero as their sentinel. Adds the Grid16x16
flipbook layout to the scripting surface.

The texture table also gains a bounded multi-name batch publish and a
replace-adopt path that keeps the prior handle. Every change bumps a timing
revision, which tells the particle pass to rebuild its timeline table.

+++ mono.engine/effects/src/Registration.cpp
@@ -149,7 +149,7 @@ namespace engine::effects {
 				writer.WriteUInt8(static_cast<uint8_t>(emitter.ShapeDirection));
 				writer.WriteUInt8(static_cast<uint8_t>(emitter.Flipbook));
 				writer.WriteUInt8(static_cast<uint8_t>(emitter.FlipbookPlayback));
-				writer.WriteUInt8(emitter.FlipbookFrames);
+				writer.WriteUInt16(emitter.FlipbookFrames);
 				writer.WriteBool(emitter.FlipbookStartRandom);
 				writer.WriteBool(emitter.LockedToPart);
 				writer.WriteBool(emitter.Enabled);
@@ -207,7 +207,7 @@ namespace engine::effects {
 				emitter.ShapeDirection = static_cast<ParticleShapeDirection>(reader.ReadUInt8());
 				emitter.Flipbook = static_cast<FlipbookLayout>(reader.ReadUInt8());
 				emitter.FlipbookPlayback = static_cast<FlipbookMode>(reader.ReadUInt8());
-				emitter.FlipbookFrames = reader.ReadUInt8();
+				emitter.FlipbookFrames = reader.ReadUInt16();
 				emitter.FlipbookStartRandom = reader.ReadBool();
 				emitter.LockedToPart = reader.ReadBool();
 				emitter.Enabled = reader.ReadBool();
